ml/embeddings/requirements.txt
beautifulsoup4
requests
langchain
openai
faiss-cpu
numpy